+++ graphrag/general/index.py
@@ -26,7 +26,7 @@ from graphrag.general.graph_extractor import DEFAULT_ENTITY_TYPES
 from graphrag.utils import graph_merge, set_entity, get_relation, set_relation, get_entity, get_graph, set_graph, \
     chunk_id, update_nodes_pagerank_nhop_neighbour
 from rag.nlp import rag_tokenizer, search
-from rag.utils.redis_conn import RedisDistributedLock
+from rag.utils.distributed_lock import RedisDistributedLock
 
 
 class Dealer:
